Removed comments and moved helper functions
